Refactored common test code
